% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{Interpolation}
\alias{Interpolation}
\title{Predict normative retina sensitivity with spacial interpolation}
\usage{
Interpolation(dt, exam = "Mesopic", ncpus = NULL, cl = NULL)
}
\arguments{
\item{dt}{A numeric matrix from SensForFit.}

\item{exam}{A string.}

\item{ncpus}{A number.}

\item{cl}{A number.}
}
\description{
Predict normative retina sensitivity with spacial interpolation
}
\examples{
# Here is an example
interpolated_pred=Interpolation(dt=ref77, ncpus=NULL, cl=NULL)
}
