% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_proc.R
\name{kriging_proc}
\alias{kriging_proc}
\title{Tune kriging model}
\usage{
kriging_proc(dt, exam)
}
\arguments{
\item{dt}{A numeric matrix from SensForFit.}

\item{exam}{A string.}
}
\description{
Tune kriging model
}
