% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCompare.R
\name{PredictCompare}
\alias{PredictCompare}
\title{Compare normative retina sensitivity prediction models}
\usage{
PredictCompare(dt, exam = "Mesopic", CalibSplit = 0.2, coverage = 0.95)
}
\arguments{
\item{dt}{A numeric matrix from SensForFit.}

\item{exam}{A string.}

\item{CalibSplit}{A number.}

\item{coverage}{A number.}
}
\description{
Compare normative retina sensitivity prediction models
}
\examples{
# Here is an example
PredictCompare(refMes, exam="Mesopic", CalibSplit=0.2, coverage=0.95)
}
