% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictNormal_rf.R
\name{PredictNormal_rf}
\alias{PredictNormal_rf}
\title{Predict normative retina sensitivity - random forest}
\usage{
PredictNormal_rf(
  dt,
  exam = "Mesopic",
  model = "LMM",
  CalibSplit = 0.2,
  coverage = 0.95
)
}
\arguments{
\item{dt}{A numeric matrix from SensForFit.}

\item{exam}{A string.}

\item{model}{A string.}

\item{CalibSplit}{A number.}

\item{coverage}{A number.}
}
\description{
Predict normative retina sensitivity - random forest
}
