% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictSensitivity.R
\name{PredictSensitivity}
\alias{PredictSensitivity}
\title{Predict normative retina sensitivity}
\usage{
PredictSensitivity(model, age, dt)
}
\arguments{
\item{model}{A string.}

\item{age}{A number}

\item{dt}{A numeric matrix from SensForFit.}
}
\description{
Predict normative retina sensitivity
}
\examples{
# Here is an example
PredictSensitivity("LMM", 77, refMes)
}
