% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ref77.R
\docType{data}
\name{ref77}
\alias{ref77}
\title{Example Dataset for a reference patient aged 77}
\format{
A data frame with 2566404 rows and 6 variables:
\describe{
\item{Examtype}{A string (Mesopic or Cyan or Red).}
\item{X_corr}{x coordinate.}
\item{Y_corr}{y coordinate.}
\item{mean}{mean sensitivity.}
\item{lwr}{lower bound of sensitivity.}
}
}
\source{
Virtual sensitivity data
}
\usage{
data(ref77)
}
\description{
Example Dataset for a reference patient aged 77
}
\keyword{datasets}
