% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refMes}
\alias{refMes}
\title{Example Dataset for a reference population}
\format{
A data frame with 7445 rows and 14 variables:
\describe{
\item{Patient}{Patient ID.}
\item{Eye}{Study eye.}
\item{Examtype}{A string (Mesopic or Cyan or Red).}
\item{Age}{Patient age.}
\item{eccentricity}{Stimulus eccentricity.}
\item{angle}{Stimulus angel.}
\item{x}{x coordinate.}
\item{y}{y coordinate.}
\item{MeanSens}{mean sensitivity.}
\item{fold}{fold for cross validation.}
}
}
\source{
Virtual sensitivity data
}
\usage{
data(refMes)
}
\description{
Example Dataset for a reference population
}
\keyword{datasets}
