% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensForFit.R
\name{SensForFit}
\alias{SensForFit}
\title{Sensitivity data pre-processing}
\usage{
SensForFit(dt, examcol, idcol, agecol, senscol, k)
}
\arguments{
\item{dt}{A numeric matrix contains all information.}

\item{examcol}{Column name for exam type.}

\item{idcol}{Column name for patient ID.}

\item{agecol}{Column name for patient age.}

\item{senscol}{Column name for outcome.}

\item{k}{A number of folds}
}
\description{
Sensitivity data pre-processing
}
\examples{
# Here is an example

}
