% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualRetina.R
\name{VisualRetina}
\alias{VisualRetina}
\title{Visualize Retina Sensitivity Data}
\usage{
VisualRetina(pred_sens, exam)
}
\arguments{
\item{pred_sens}{A data frame containing predicted sensitivity data from interpolation or prediction.}

\item{exam}{Exam type.}
}
\description{
Visualize Retina Sensitivity Data
}
