% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{Interpolation}
\alias{Interpolation}
\title{Predict normative retina sensitivity}
\usage{
Interpolation(dt, exam, model)
}
\arguments{
\item{dt}{A numeric matrix from SensForFit.}

\item{exam}{A string.}

\item{model}{A string.}
}
\description{
Predict normative retina sensitivity
}
\examples{
# Here is an example

}
