% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictNormal.R
\name{PredictNormal}
\alias{PredictNormal}
\title{Predict normative retina sensitivity}
\usage{
PredictNormal(dt, exam = "Mesopic", model = "LMM")
}
\arguments{
\item{dt}{A numeric matrix from SensForFit.}

\item{exam}{A string.}

\item{model}{A string.}
}
\description{
Predict normative retina sensitivity
}
\examples{
# Here is an example

}
