% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref77}
\alias{ref77}
\title{Example Dataset for a reference patient aged 77}
\format{
A data frame with 2566404 rows and 6 variables:
\describe{
\item{Examtype}{A string (Mesopic or Cyan or Red).}
\item{eccentricity}{Stimulus eccentricity.}
\item{angle}{Stimulus angel.}
\item{x}{x coordinate.}
\item{y}{y coordinate.}
\item{MeanSens}{mean sensitivity.}
}
}
\source{
Virtual sensitivity data
}
\usage{
data(ref77)
}
\description{
Example Dataset for a reference patient aged 77
}
\keyword{datasets}
